Creates a series of crypto utilities
This commit creates a series of 3 cryptography utilities used to
encrypt the data that will be stored by the application as part of
its content.

Everytime we want to store an information, the idea is to generate
a random encryption key that will be stored part of the SecureStore
while the encrypted information using this generated key will be
stored in the regular FileSystem.

This solution will allow us to not be blocked by the low capacity
of the SecureStore allowed on iOS and Android.

Part of this commit, we find the `generateEncryptionKey` utility
creating a random encryption key having the requested amount of
chars.

We also have an `encrypt` utility randomly generating a key and
using the generated key to encrypt the requested message.

And finally we have a `decrypt` key using the provided key to
decrypt a provided message.

These utilities are based on the work made by @neverdull-agency
to implement an unlimited secured expo storage, following the
message by @Cretezy in the issue
Cretezy/redux-persist-expo-securestore#9.