Create tests for crypto utilities
This commit adds a series of test cases to make sure the crypto
utilities continue to work when updated.

A first test sets a series of variables for these test cases with
the value of a first encryption.

A second test tries to decrypt this first encryption using the
generated key and the generated encrypted string.

A third test makes sure that we can only decrypt an encrypted
string using the key generated during the encryption.

A last test makes sure that the random generation of encryption
keys is really random by generating three different keys.