Impelement a set of storage utilities
This set of utilities will allow us to easily store encrypted info
in the FileSystem of the device while using an AES based encryption
per value, with the key used for it stored in the SecureStorage of
the OS.

The `createDirectory` utility is creating the folder that will host
our store if not existing.

The `fixedStorageUri` makes sure that we are always creating files
for our store part of our directory.

The `generateStorageUri` utility generates the file URI for our
store.

The `getAsync` utility fetches an information and decrypt it from
our current store.

The `removeAsync` utility removes a requested information from the
current store.

The `setAsync` utility adds a new information and encrypt it in our
store.

Finally, the `writeStore` utility is used to centralize all actions
used to write the updated Store in a new file (name randomly
generated following uuid v4) and to store the new URI in the
SecureStore of the OS.